type for gcc on osx
git-svn-id: http://patchdiff2.googlecode.com/svn/trunk@29 b4ccacd4-e174-7a79-44b8-234ebb245366

+++ hash.cpp
@@ -28,7 +28,7 @@
 /* description: Initializes hash table to NULL    */
 /*------------------------------------------------*/
 
-hpsig_t * hash_init(unsigned int num)
+hpsig_t * hash_init(size_t num)
 {
 	unsigned int i;
 	hpsig_t * hsig;
